Added DataManager class
The DataManager class will be the one that player/author call to set or
get worldData. The purpose for creating this class was 1) to have an in
memory map of the current world data, 2) avoid the JSON to java
conversion every time world data is called, 3) It also allows the
flexibility to use the worldData without storing it to a file.

Added GsonBuilder to create GSON object which allows for the prettyprint
method. This allows the JSON text to be formatted more nicely instead of
everything being output in one line.